+++ src/main/java/subway/dto/StationResponse.java
@@ -1,8 +1,8 @@
-package subway;
+package subway.dto;
 
 public class StationResponse {
-    private Long id;
-    private String name;
+    private final Long id;
+    private final String name;
 
     public StationResponse(Long id, String name) {
         this.id = id;
